+++ libmkpass/passphrase_patterns.h
@@ -1,10 +1,11 @@
 #pragma once
 
+#include <cstddef>
 #include <vector>
 #include "word_classes.h"
 
 using PassphrasePattern = std::vector<WordClasses>;
 using PatternsList = std::vector<PassphrasePattern>;
 
-PatternsList GetPassphrasePatterns(int length);
-int GetMaxPassphrasePatternLength();
+PatternsList GetPassphrasePatterns(size_t length);
+size_t GetMaxPassphrasePatternLength();
